Issue #4586: Moved resources for AtclauseOrder check test
