+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/AtclauseOrderCheckTest.java
@@ -36,7 +36,8 @@ public class AtclauseOrderCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "javadoc" + File.separator + filename);
+                + "javadoc" + File.separator
+                + "atclauseorder" + File.separator + filename);
     }
 
     @Test
@@ -60,7 +61,7 @@ public class AtclauseOrderCheckTest extends BaseCheckTestSupport {
         final DefaultConfiguration checkConfig = createCheckConfig(AtclauseOrderCheck.class);
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
 
-        verify(checkConfig, getPath("InputCorrectAtClauseOrder.java"), expected);
+        verify(checkConfig, getPath("InputAtClauseOrderCorrect.java"), expected);
     }
 
     @Test
@@ -110,7 +111,7 @@ public class AtclauseOrderCheckTest extends BaseCheckTestSupport {
             "278: " + getCheckMessage(MSG_KEY, tagOrder),
             "288: " + getCheckMessage(MSG_KEY, tagOrder),
         };
-        verify(checkConfig, getPath("InputIncorrectAtClauseOrder.java"), expected);
+        verify(checkConfig, getPath("InputAtClauseOrderIncorrect.java"), expected);
     }
 
     @Test
@@ -127,6 +128,6 @@ public class AtclauseOrderCheckTest extends BaseCheckTestSupport {
         final String[] expected = {
             "113: " + getCheckMessage(MSG_KEY, tagOrder),
         };
-        verify(checkConfig, getPath("InputIncorrectAtClauseOrder.java"), expected);
+        verify(checkConfig, getPath("InputAtClauseOrderIncorrect.java"), expected);
     }
 }
