+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/atclauseorder/InputAtClauseOrderCorrect.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc;
+package com.puppycrawl.tools.checkstyle.checks.javadoc.atclauseorder;
 
 import java.io.Serializable;
